seperated code in hpp to cpp file
