+++ src/builders/JSBuilder.hpp
@@ -20,14 +20,7 @@
 class JSBuilder : public ContentBuilder {
 public:
 
-    JSBuilder(const std::string& path, const std::string& serverRoot) : ContentBuilder(path, serverRoot) {
-
-        json = getJSONFromFile(serverRoot + "/files_maps/js_file_map.json");
-
-        pageName = getFileNameWithoutExtension(path);
-
-        builJS();
-    }
+    JSBuilder(const std::string& path, const std::string& serverRoot);
 
 private:
 
@@ -35,37 +28,9 @@ private:
 
     std::string pageName;
 
-    static std::string getFileNameWithoutExtension(const std::string& path) {
-        // Find the last position of the slash or backslash
-        size_t lastSlashPos = path.find_last_of('/');
+    static std::string getFileNameWithoutExtension(const std::string& path);
 
-        // Get the file name from the path
-        std::string fileName = (lastSlashPos == std::string::npos) ? path : path.substr(lastSlashPos + 1);
-
-        // Find the last position of the dot
-        size_t lastDotPos = fileName.find_last_of('.');
-        // Remove the extension from the file name
-        if (lastDotPos != std::string::npos) {
-            fileName = fileName.substr(0, lastDotPos);
-        }
-
-        return fileName;
-    }
-
-    void builJS() {
-
-        JSONParser jsonForFile = json->getObject(pageName);
-
-        // If there are no files to include, return
-        if(jsonForFile.getKeys().empty()) return;
-
-        builtFile = "";
-
-        // Build the file by including all the files
-        for (const auto& key : jsonForFile.getKeys()) {
-            builtFile += loadFile(root + "/assets/js" + jsonForFile.getString(key)) + "\n\n";
-        }
-    }
+    void builJS();
 };
 
 #endif //JSBUILDER_HPP
